st/nine: Explicit nine requirements
This patches raises nine requirements and disables nine for old
hw that don't have match them.

It would be possible to make a lot of things work with these hw,
though not everything, but it needs special care for them in the
code, and since they are very old, it's better to drop explicitly
support for them.

Before this patch some things would have worked for these hw,
but users would face a lot of things not working.

Signed-off-by: Axel Davy <axel.davy@ens.fr>